requirements.txt
aiohttp
aiosignal
APScheduler
asgiref
async-timeout
attrs
certifi
charset-normalizer
coreapi
coreschema
dateparser
dj-database-url
Django
django-4-jet
django-cors-headers
django-extensions
django-filter
djangorestframework
drf-yasg
environs
frozenlist
gunicorn
idna
inflection
itypes
Jinja2
Markdown
MarkupSafe
marshmallow
multidict
packaging
psycopg2-binary
pyparsing
python-binance
python-dateutil
python-dotenv
pytz
pytz-deprecation-shim
regex
requests
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
tzdata
tzlocal
ujson
uritemplate
urllib3
websockets
yarl
pillow
django_cleanup
dj-rest-auth
django-allauth
daphne
channels
channels-redis
solana
httpx
apscheduler